refactor: moved handlers + made new handlers
