+++ src/main/java/roomescape/model/errors/ReservationNotFoundException.java
@@ -1,0 +1,3 @@
+package roomescape.model.errors;
+
+public class ReservationNotFoundException extends IllegalArgumentException{}
