+++ src/main/java/roomescape/model/handler/DomainExceptionHandler.java
@@ -1,4 +1,4 @@
-package roomescape.common;
+package roomescape.model.handler;
 
 import org.springframework.http.HttpStatus;
 import org.springframework.validation.FieldError;
@@ -6,21 +6,16 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import roomescape.model.errors.ReservationNotFoundException;
 
 import java.util.HashMap;
 import java.util.Map;
 
 @RestControllerAdvice
-public class IncorrectAttributeErrorHandler {
-    @ExceptionHandler(MethodArgumentNotValidException.class)
+public class DomainExceptionHandler {
+    @ExceptionHandler(ReservationNotFoundException.class)
     @ResponseStatus(HttpStatus.BAD_REQUEST)
-    public Map<String, String> handleIncorrectAttributes(MethodArgumentNotValidException exception) {
-        Map<String, String> errors = new HashMap<>();
-
-        for (FieldError error : exception.getBindingResult().getFieldErrors()) {
-            errors.put(error.getField(), error.getDefaultMessage());
-        }
-
-        return errors;
+    public String handleReservationNotFoundError(ReservationNotFoundException exception) {
+        return "없는 예약번호 입니다.";
     }
 }
